backend/requirements.txt
# --- Core Backend ---
git+https://github.com/tiangolo/fastapi.git@master  # FastAPI pre-release supporting Pydantic v2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.3

# --- Database ---
sqlalchemy==2.0.29

# --- Chat & Realtime ---
websockets==12.0
aiofiles==23.2.1
aioredis==2.0.1

# --- Data & ML ---
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
faiss-cpu==1.7.4
sentence-transformers==2.6.1

# --- LangChain & Community ---
langchain==0.1.13
langchain-community==0.0.30

# --- Vector DB (Optional) ---
pinecone-client==3.2.2  # remove if unused

# --- Hugging Face APIs ---
huggingface-hub==0.22.2
transformers==4.39.3
accelerate==0.28.0
torch==2.2.2

# --- Reports / File Handling ---
reportlab==4.1.0         # PDF generation
python-multipart==0.0.9  # Needed if you accept file uploads

# --- Utilities ---
tqdm==4.66.2

# --- Authentication ---
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
